YieldNexus: Rename adapter's name (#3508)
* Add support for multiple media types. Add test coverage.

* Add support for multiple media types. Add test coverage.

* Modify multi-format ads handler. Modify tests.

* Rename yieldnexus bid adapter to fix download issue
